style(wbi): 格式化 WBI 密钥索引表数组结构
- 将多行数组定义改为单行格式，提高代码可读性
- 移除多余的 toShort() 类型转换调用

style(websocket): 简化数据包编解码器中的类型转换

- 移除 HEADER_SIZE 的冗余类型转换
- 保持代码简洁性和一致性

+++ fanclub-bilibili/src/main/kotlin/websocket/PacketCodec.kt
@@ -11,7 +11,7 @@ private const val HEADER_SIZE = 16
 fun makePacket(data: ByteArray, operation: WsOperation): ByteString {
     val buffer = Buffer()
     buffer.writeInt(HEADER_SIZE + data.size)
-    buffer.writeShort(HEADER_SIZE.toShort())
+    buffer.writeShort(HEADER_SIZE)
     buffer.writeShort(1)
     buffer.writeInt(operation.value)
     buffer.writeInt(1)
